Required ident to server to make connection added.
Added a security feature to the VTX server. It now only accepts connections only from valid VTX clients. Once the client makes a websocket connection to the VTX server, it will be asked to ident (CSI 0 c). If it does not get a CSI ?50;86;84;88 c back in 30 seconds, the server disconnects the connection. If it performs a successful handshake, the connection to the node BBS will be made. This is to keep fTelnet's, hackers, and other riffraff out of a VTX server.